fix: update i18n plugin to work
