+++ cypress/component/advanced/i18n/README.md
@@ -1,0 +1,9 @@
+# i18n example
+
+Using [vue-i18n](https://kazupon.github.io/vue-i18n/) and [@intlify/vue-i18n-loader](https://github.com/intlify/vue-i18n-loader/tree/master) stable.
+
+See [TranslatedMessage.vue](TranslatedMessage.vue) and its test in [spec.js](spec.js)
+
+![i18n test in action](./images/i18n.gif)
+
+**Note:** the `vue-i18n-loader` requires a webpack rule, see Webpack config in the root folder.
