+++ package.json
@@ -88,7 +88,7 @@
     "@babel/plugin-transform-modules-commonjs": "7.10.4",
     "@cypress/code-coverage": "3.8.1",
     "@cypress/webpack-preprocessor": "5.4.1",
-    "@kazupon/vue-i18n-loader": "0.5.0",
+    "@intlify/vue-i18n-loader": "1.0.0",
     "babel-plugin-istanbul": "6.0.0",
     "common-tags": "1.8.0",
     "debug": "4.1.1",
